fix: migrate to an animation file, update cheatsheet readme
